SciDAVis 2.9 release

+++ mkMacDist.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 # Create a distributable installable package
 
-MAC_DIST_DIR=scidavis/scidavis.app/Contents/MacOS
-RES_DIR=scidavis/scidavis.app/Contents/Resources
+if [ $# -gt 1 ]; then
+    BUNDLE=$1
+else
+    BUNDLE=scidavis.app
+fi
+
+MAC_DIST_DIR=$BUNDLE/Contents/MacOS
+RES_DIR=$BUNDLE/Contents/Resources
 version=`grep scidavis_version libscidavis/src/version.cpp|tail -1|cut -f5 -d' '|tr -d '";'`
 if [ $version = '"unknown"' ]; then
     version=0.0.0.0
@@ -14,13 +20,13 @@ rewrite_dylibs()
     echo "rewrite_dylibs $target"
     otool -L $target|grep opt/local|cut -f1 -d' '|while read dylib; do
         # avoid infinite loops
-        if [ "${dylib##*/}" != "${target##*/}" ]; then 
+        if [ "${dylib##*/}" != "${target##*/}" -a ! -f $MAC_DIST_DIR/${dylib##*/} ]; then 
             cp -f $dylib $MAC_DIST_DIR
             chmod u+rw $MAC_DIST_DIR/${dylib##*/}
             rewrite_dylibs $MAC_DIST_DIR/${dylib##*/}
-            echo "install_name_tool -change $dylib @executable_path/${dylib##*/} $target"
         fi
         install_name_tool -change $dylib @executable_path/${dylib##*/} $target
+        install_name_tool -id @executable_path/${dylib##*/} $target
     done
     otool -L $target|grep usr/local|cut -f1 -d' '|while read dylib; do
         # avoid infinite loops
@@ -33,31 +39,27 @@ rewrite_dylibs()
     done
 }
 
+mkdir -p $MAC_DIST_DIR
 cp scidavis/scidavis $MAC_DIST_DIR
-rewrite_dylibs $MAC_DIST_DIR/scidavis
-
 chmod u+w $MAC_DIST_DIR/*
 
-# Generic resources required for Qt
-cp -rf /opt/local/libexec/qt5/Library/Frameworks/QtGui.framework/Resources/qt_menu.nib $RES_DIR
 
-# python resources
+## python resources
 mkdir -p $RES_DIR/lib
-cp -rf /opt/local/Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9 $RES_DIR/lib
+cp -rf /opt/local/Library/Frameworks/Python.framework/Versions/Current/lib/python* $RES_DIR/lib
 
 # python resources contain some dynamic libraries that need rewriting
 find $RES_DIR/lib -name "*.so" -print | while read soname; do
-    otool -L $soname|grep /opt/local|cut -f1 -d' '|while read oldName; do
-        install_name_tool -change $oldName @executable_path/${oldName##*/} $soname
-        done
+    rewrite_dylibs $soname
 done
 
 # copy translation files
 cp -rf scidavis/translations $MAC_DIST_DIR
 
 # copy icon, and create mainfest
+mkdir -p $RES_DIR
 cp -f scidavis/icons/scidavis.icns $RES_DIR
-cat >scidavis/scidavis.app/Contents/Info.plist <<EOF
+cat >$BUNDLE/Contents/Info.plist <<EOF
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN"
  "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
@@ -86,14 +88,19 @@ cat >scidavis/scidavis.app/Contents/Info.plist <<EOF
 </plist>
 EOF
 
-codesign -s "Developer ID Application" --deep --force scidavis/scidavis.app
+macdeployqt $BUNDLE
+# macdeployqt fails to fix up python references
+rewrite_dylibs $MAC_DIST_DIR/scidavis
+
+find $BUNDLE/Contents/Resources \( -name "*.dylib" -o -name "*.so" -o -name "*a" \) -exec codesign -s "Developer ID Application" --options runtime --timestamp --force {} \; -print
+codesign -s "Developer ID Application" --options runtime --timestamp --deep --force $BUNDLE
 if [ $? -ne 0 ]; then
     echo "try running this script on the Mac GUI desktop, not ssh shell"
-    exit
+    exit 1
 fi
 
 rm -f scidavis-$version-mac-dist.dmg
-hdiutil create -srcfolder scidavis/scidavis.app -volname SciDAVis -fs HFS+ -fsargs "-c c=64,a=16,e=16" -format UDRW -size 200M temp.dmg
+hdiutil create -srcfolder $BUNDLE -volname SciDAVis -fs HFS+ -fsargs "-c c=64,a=16,e=16" -format UDRW -size 500M temp.dmg
 hdiutil convert -format UDZO -imagekey zlib-level=9 -o scidavis-$version-mac-dist.dmg temp.dmg 
 rm -f temp.dmg
 codesign -s "Developer ID Application" scidavis-$version-mac-dist.dmg
